renamed to locates
