+++ cmd/projects/locates/cmd_locates_test.go
@@ -1,4 +1,4 @@
-package locales
+package locates
 
 import (
 	"bytes"
